refactor: 더 직관적이게 valueOf 함수 로직을 수정
